refactor: split character.go into focused files for better organization
- Moved combat methods to character_combat.go (Attack, weapon proficiency, fighting styles)
- Moved proficiency methods to character_proficiencies.go (saving throws, skill checks)
- Moved effect management to character_effects.go (status effects, effect sync)
- Added character_resources.go for resource management
- Reduced character.go from 1631 to ~1200 lines
- All tests passing, no functionality changes

This refactoring improves code organization and prepares for future migration to rpg-toolkit project.
