+++ internal/entities/character.go
@@ -8,8 +8,6 @@ import (
 
 	"github.com/KirkDiggler/dnd-bot-discord/internal/dice"
 	"github.com/KirkDiggler/dnd-bot-discord/internal/effects"
-	"github.com/KirkDiggler/dnd-bot-discord/internal/entities/attack"
-	"github.com/KirkDiggler/dnd-bot-discord/internal/entities/damage"
 )
 
 type Slot string
@@ -97,412 +95,6 @@ func (c *Character) getDiceRoller() dice.Roller {
 		c.diceRoller = dice.NewRandomRoller()
 	}
 	return c.diceRoller
-}
-
-func (c *Character) Attack() ([]*attack.Result, error) {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-
-	if c.EquippedSlots == nil {
-		// Improvised weapon range or melee
-		// No equipped slots, using improvised melee
-		a, err := c.improvisedMelee()
-		if err != nil {
-			return nil, err
-		}
-
-		return []*attack.Result{
-			a,
-		}, nil
-
-	}
-
-	// Check main hand slot
-	if c.EquippedSlots[SlotMainHand] != nil {
-		if weap, ok := c.EquippedSlots[SlotMainHand].(*Weapon); ok {
-			attacks := make([]*attack.Result, 0)
-
-			// Check proficiency while we have the mutex
-			directProf := c.hasWeaponProficiencyInternal(weap.GetKey())
-			categoryProf := c.hasWeaponCategoryProficiency(weap.WeaponCategory)
-			isProficient := directProf || categoryProf
-			// Proficiency check completed
-
-			// Calculate ability bonus based on weapon type
-			var abilityBonus int
-			if c.Attributes != nil {
-				switch weap.WeaponRange {
-				case "Ranged":
-					if c.Attributes[AttributeDexterity] != nil {
-						abilityBonus = c.Attributes[AttributeDexterity].Bonus
-					}
-				case "Melee":
-					if c.Attributes[AttributeStrength] != nil {
-						abilityBonus = c.Attributes[AttributeStrength].Bonus
-					}
-				}
-			}
-
-			// Calculate proficiency bonus if proficient
-			proficiencyBonus := 0
-			if isProficient {
-				proficiencyBonus = 2 + ((c.Level - 1) / 4)
-			}
-
-			attackBonus := abilityBonus + proficiencyBonus
-			damageBonus := abilityBonus // Base damage bonus from ability modifier
-
-			// Apply fighting style bonuses
-			attackBonus, damageBonus = c.applyFightingStyleBonuses(weap, attackBonus, damageBonus)
-
-			// Apply damage bonuses from active effects (e.g., rage)
-			// Use the weapon's actual range type
-			attackType := strings.ToLower(weap.WeaponRange)
-			var err error
-			damageBonus, err = c.applyActiveEffectDamageBonus(damageBonus, attackType)
-			if err != nil {
-				log.Printf("ERROR: Failed to apply active effect damage bonus: %v", err)
-				// Continue with base damage bonus
-			}
-
-			log.Printf("Final attack bonus: +%d (ability: %d, proficiency: %d)", attackBonus, abilityBonus, proficiencyBonus)
-			log.Printf("Final damage bonus: +%d", damageBonus)
-
-			// Roll the attack with fighting style consideration
-			var attak1 *attack.Result
-			fightingStyle := c.getFightingStyle()
-			// Great Weapon Fighting only applies to two-handed melee weapons
-			useGreatWeaponFighting := fightingStyle == "great_weapon" && weap.IsTwoHanded() && weap.IsMelee()
-
-			if weap.IsTwoHanded() && weap.TwoHandedDamage != nil {
-				if useGreatWeaponFighting {
-					attak1, err = attack.RollAttackWithFightingStyle(c.getDiceRoller(), attackBonus, damageBonus, weap.TwoHandedDamage, "great_weapon")
-				} else {
-					attak1, err = attack.RollAttack(c.getDiceRoller(), attackBonus, damageBonus, weap.TwoHandedDamage)
-				}
-			} else {
-				attak1, err = attack.RollAttack(c.getDiceRoller(), attackBonus, damageBonus, weap.Damage)
-			}
-
-			if err != nil {
-				log.Printf("Weapon attack error: %v", err)
-				return nil, err
-			}
-			log.Printf("Weapon attack successful")
-			attacks = append(attacks, attak1)
-
-			if c.EquippedSlots[SlotOffHand] != nil {
-				if offWeap, offOk := c.EquippedSlots[SlotOffHand].(*Weapon); offOk {
-					// Same process for off-hand weapon
-					offHandProficient := c.hasWeaponProficiencyInternal(offWeap.GetKey()) ||
-						c.hasWeaponCategoryProficiency(offWeap.WeaponCategory)
-
-					var offHandAbilityBonus int
-					switch offWeap.WeaponRange {
-					case "Ranged":
-						offHandAbilityBonus = c.Attributes[AttributeDexterity].Bonus
-					case "Melee":
-						offHandAbilityBonus = c.Attributes[AttributeStrength].Bonus
-					}
-
-					offHandProficiencyBonus := 0
-					if offHandProficient {
-						offHandProficiencyBonus = 2 + ((c.Level - 1) / 4)
-					}
-
-					offHandAttackBonus := offHandAbilityBonus + offHandProficiencyBonus
-					offHandDamageBonus := offHandAbilityBonus
-
-					// Apply fighting style bonuses to off-hand
-					offHandAttackBonus, offHandDamageBonus = c.applyFightingStyleBonusesWithHand(offWeap, offHandAttackBonus, offHandDamageBonus, SlotOffHand)
-
-					// Apply damage bonuses from active effects (e.g., rage) to off-hand
-					offHandDamageBonus, err = c.applyActiveEffectDamageBonus(offHandDamageBonus, "melee")
-					if err != nil {
-						log.Printf("ERROR: Failed to apply active effect damage bonus to off-hand: %v", err)
-						// Continue with current bonus
-					}
-
-					attak2, err := attack.RollAttack(c.getDiceRoller(), offHandAttackBonus, offHandDamageBonus, offWeap.Damage)
-					if err != nil {
-						return nil, err
-					}
-					attacks = append(attacks, attak2)
-				}
-			}
-
-			log.Printf("Returning %d attack results", len(attacks))
-			return attacks, nil
-		} else {
-			log.Printf("Main hand equipment is not a weapon: %T", c.EquippedSlots[SlotMainHand])
-		}
-	} else {
-		log.Printf("No main hand equipment")
-	}
-
-	if c.EquippedSlots[SlotTwoHanded] != nil {
-		log.Printf("Checking two-handed slot...")
-		log.Printf("Two-handed slot type: %T", c.EquippedSlots[SlotTwoHanded])
-		if weap, ok := c.EquippedSlots[SlotTwoHanded].(*Weapon); ok {
-			log.Printf("Two-handed weapon found: %s", weap.GetName())
-			// Check proficiency while we have the mutex
-			directProf := c.hasWeaponProficiencyInternal(weap.GetKey())
-			categoryProf := c.hasWeaponCategoryProficiency(weap.WeaponCategory)
-			isProficient := directProf || categoryProf
-			// Proficiency check completed
-
-			// Calculate ability bonus based on weapon type
-			var abilityBonus int
-			if c.Attributes != nil {
-				switch weap.WeaponRange {
-				case "Ranged":
-					if c.Attributes[AttributeDexterity] != nil {
-						abilityBonus = c.Attributes[AttributeDexterity].Bonus
-					}
-				case "Melee":
-					if c.Attributes[AttributeStrength] != nil {
-						abilityBonus = c.Attributes[AttributeStrength].Bonus
-					}
-				}
-			}
-
-			// Calculate proficiency bonus if proficient
-			proficiencyBonus := 0
-			if isProficient {
-				proficiencyBonus = 2 + ((c.Level - 1) / 4)
-			}
-
-			attackBonus := abilityBonus + proficiencyBonus
-			damageBonus := abilityBonus
-
-			// Apply fighting style bonuses
-			attackBonus, damageBonus = c.applyFightingStyleBonuses(weap, attackBonus, damageBonus)
-
-			// Apply damage bonuses from active effects (e.g., rage)
-			// Use the weapon's actual range type
-			attackType := strings.ToLower(weap.WeaponRange)
-			var err error
-			damageBonus, err = c.applyActiveEffectDamageBonus(damageBonus, attackType)
-			if err != nil {
-				log.Printf("ERROR: Failed to apply active effect damage bonus: %v", err)
-				// Continue with base damage bonus
-			}
-
-			log.Printf("Final attack bonus: +%d (ability: %d, proficiency: %d)", attackBonus, abilityBonus, proficiencyBonus)
-			log.Printf("Final damage bonus: +%d", damageBonus)
-
-			// Two-handed weapons often have special damage
-			var dmg *damage.Damage
-			if weap.TwoHandedDamage != nil {
-				dmg = weap.TwoHandedDamage
-			} else {
-				dmg = weap.Damage
-			}
-
-			// Apply Great Weapon Fighting for two-handed melee weapons
-			fightingStyle := c.getFightingStyle()
-			var a *attack.Result
-			if fightingStyle == "great_weapon" && weap.IsMelee() {
-				a, err = attack.RollAttackWithFightingStyle(c.getDiceRoller(), attackBonus, damageBonus, dmg, "great_weapon")
-			} else {
-				a, err = attack.RollAttack(c.getDiceRoller(), attackBonus, damageBonus, dmg)
-			}
-			if err != nil {
-				return nil, err
-			}
-
-			return []*attack.Result{
-				a,
-			}, nil
-		}
-	}
-
-	a, err := c.improvisedMelee()
-	if err != nil {
-		return nil, err
-	}
-
-	return []*attack.Result{
-		a,
-	}, nil
-}
-
-// HasWeaponProficiency checks if the character is proficient with a weapon (thread-safe)
-func (c *Character) HasWeaponProficiency(weaponKey string) bool {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-
-	// First check direct weapon proficiency
-	if c.hasWeaponProficiencyInternal(weaponKey) {
-		return true
-	}
-
-	// Check if we have the weapon to get its category
-	weapon := c.getEquipment(weaponKey)
-	if weapon != nil {
-		if w, ok := weapon.(*Weapon); ok && w.WeaponCategory != "" {
-			return c.hasWeaponCategoryProficiency(w.WeaponCategory)
-		}
-	}
-
-	// Also check equipped weapons
-	for _, equipped := range c.EquippedSlots {
-		if equipped != nil && equipped.GetKey() == weaponKey {
-			if w, ok := equipped.(*Weapon); ok && w.WeaponCategory != "" {
-				return c.hasWeaponCategoryProficiency(w.WeaponCategory)
-			}
-		}
-	}
-
-	return false
-}
-
-// hasWeaponProficiencyInternal checks proficiency without locking (must be called with mutex held)
-func (c *Character) hasWeaponProficiencyInternal(weaponKey string) bool {
-	if c.Proficiencies == nil {
-		return false
-	}
-
-	weaponProficiencies, exists := c.Proficiencies[ProficiencyTypeWeapon]
-	if !exists {
-		return false
-	}
-
-	for _, prof := range weaponProficiencies {
-		if prof.Key == weaponKey {
-			return true
-		}
-	}
-
-	return false
-}
-
-// hasWeaponCategoryProficiency checks if the character has proficiency with a weapon category
-// This handles proficiencies like "simple-weapons" or "martial-weapons"
-func (c *Character) hasWeaponCategoryProficiency(weaponCategory string) bool {
-	if c.Proficiencies == nil || weaponCategory == "" {
-		return false
-	}
-
-	weaponProficiencies, exists := c.Proficiencies[ProficiencyTypeWeapon]
-	if !exists {
-		return false
-	}
-
-	// Map weapon categories to proficiency keys (case-insensitive)
-	categoryMap := map[string]string{
-		"simple":  "simple-weapons",
-		"martial": "martial-weapons",
-	}
-
-	// Convert to lowercase for case-insensitive comparison
-	lowerCategory := strings.ToLower(weaponCategory)
-	profKey, exists := categoryMap[lowerCategory]
-	if !exists {
-		return false
-	}
-
-	for _, prof := range weaponProficiencies {
-		if prof.Key == profKey {
-			return true
-		}
-	}
-
-	return false
-}
-
-func (c *Character) improvisedMelee() (*attack.Result, error) {
-	// Check if character has Martial Arts feature (monks)
-	hasMartialArts := false
-	for _, feature := range c.Features {
-		if feature != nil && feature.Key == "martial-arts" {
-			hasMartialArts = true
-			break
-		}
-	}
-
-	// Determine ability bonus - monks can use DEX instead of STR
-	bonus := 0
-	if hasMartialArts && c.Attributes != nil {
-		// Use the higher of STR or DEX for monks
-		strBonus := 0
-		dexBonus := 0
-		if c.Attributes[AttributeStrength] != nil {
-			strBonus = c.Attributes[AttributeStrength].Bonus
-		}
-		if c.Attributes[AttributeDexterity] != nil {
-			dexBonus = c.Attributes[AttributeDexterity].Bonus
-		}
-		bonus = strBonus
-		if dexBonus > strBonus {
-			bonus = dexBonus
-		}
-	} else if c.Attributes != nil && c.Attributes[AttributeStrength] != nil {
-		// Non-monks use STR
-		bonus = c.Attributes[AttributeStrength].Bonus
-	}
-
-	// Apply damage bonuses from active effects (e.g., rage) to improvised attacks
-	damageBonus, err := c.applyActiveEffectDamageBonus(bonus, "melee")
-	if err != nil {
-		log.Printf("ERROR: Failed to apply active effect damage bonus to improvised: %v", err)
-		damageBonus = bonus // Fall back to base
-	}
-
-	attackResult, err := c.getDiceRoller().Roll(1, 20, 0)
-	if err != nil {
-		return nil, err
-	}
-	damageResult, err := c.getDiceRoller().Roll(1, 4, 0)
-	if err != nil {
-		return nil, err
-	}
-
-	return &attack.Result{
-		AttackRoll:   attackResult.Total + bonus,
-		DamageRoll:   damageResult.Total + damageBonus,
-		AttackType:   damage.TypeBludgeoning,
-		AttackResult: attackResult,
-		DamageResult: damageResult,
-		WeaponDamage: &damage.Damage{
-			DiceCount:  1,
-			DiceSize:   4,
-			Bonus:      0,
-			DamageType: damage.TypeBludgeoning,
-		},
-	}, nil
-}
-
-// applyActiveEffectDamageBonus applies damage bonuses from active effects like rage
-// Returns the modified damage and any error encountered
-func (c *Character) applyActiveEffectDamageBonus(baseDamage int, damageType string) (int, error) {
-	// Get damage modifiers from the new status effect system
-	conditions := map[string]string{
-		"attack_type": damageType,
-	}
-
-	modifiers := c.getDamageModifiersInternal(conditions)
-	effectBonus := 0
-
-	for _, mod := range modifiers {
-		// Parse modifier value (e.g., "+2", "+3", "+4")
-		if mod.Value != "" && mod.Value[0] == '+' {
-			var parsedBonus int
-			if n, err := fmt.Sscanf(mod.Value, "+%d", &parsedBonus); err == nil && n == 1 {
-				effectBonus += parsedBonus
-			}
-		}
-	}
-
-	baseDamage += effectBonus
-
-	// Also check the old system for backward compatibility
-	if c.Resources != nil {
-		oldEffectBonus := c.Resources.GetTotalDamageBonus(damageType)
-		baseDamage += oldEffectBonus
-	}
-
-	return baseDamage, nil
 }
 
 func (c *Character) getEquipment(key string) Equipment {
@@ -630,45 +222,6 @@ func (c *Character) SetHitpoints() {
 
 	c.MaxHitPoints = c.HitDie + c.Attributes[AttributeConstitution].Bonus
 	c.CurrentHitPoints = c.MaxHitPoints
-}
-
-// GetResources returns the character's resources, initializing if needed
-func (c *Character) GetResources() *CharacterResources {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-
-	if c.Resources == nil {
-		c.initializeResourcesInternal()
-	}
-	return c.Resources
-}
-
-// InitializeResources sets up the character's resources based on class and level
-func (c *Character) InitializeResources() {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	c.initializeResourcesInternal()
-}
-
-// initializeResourcesInternal is the internal resource initialization (caller must hold lock)
-func (c *Character) initializeResourcesInternal() {
-	if c.Resources == nil {
-		c.Resources = &CharacterResources{}
-	}
-
-	// Initialize basic resources
-	if c.Class != nil {
-		c.Resources.Initialize(c.Class, c.Level)
-	}
-
-	// Set HP based on character's max HP (includes CON bonus)
-	c.Resources.HP = HPResource{
-		Current: c.CurrentHitPoints,
-		Max:     c.MaxHitPoints,
-	}
-
-	// Initialize class-specific abilities at level 1
-	c.initializeClassAbilities()
 }
 
 // getCharismaModifier returns the character's Charisma modifier
@@ -823,40 +376,6 @@ func (c *Character) AddInventory(e Equipment) {
 
 	c.Inventory[e.GetEquipmentType()] = append(c.Inventory[e.GetEquipmentType()], e)
 	c.mu.Unlock()
-}
-
-func (c *Character) AddProficiency(p *Proficiency) {
-	if c.Proficiencies == nil {
-		c.Proficiencies = make(map[ProficiencyType][]*Proficiency)
-	}
-	c.mu.Lock()
-	defer c.mu.Unlock()
-
-	if c.Proficiencies[p.Type] == nil {
-		c.Proficiencies[p.Type] = make([]*Proficiency, 0)
-	}
-
-	// Check for duplicates
-	for _, existing := range c.Proficiencies[p.Type] {
-		if existing.Key == p.Key {
-			return // Already have this proficiency
-		}
-	}
-
-	c.Proficiencies[p.Type] = append(c.Proficiencies[p.Type], p)
-}
-
-// SetProficiencies replaces all proficiencies of a given type
-// This is used when selecting proficiencies during character creation
-func (c *Character) SetProficiencies(profType ProficiencyType, proficiencies []*Proficiency) {
-	if c.Proficiencies == nil {
-		c.Proficiencies = make(map[ProficiencyType][]*Proficiency)
-	}
-	c.mu.Lock()
-	defer c.mu.Unlock()
-
-	// Replace all proficiencies of this type
-	c.Proficiencies[profType] = proficiencies
 }
 
 func (c *Character) AddAbilityScoreBonus(attr Attribute, bonus int) {
@@ -1180,97 +699,6 @@ func (c *Character) Clone() *Character {
 	return clone
 }
 
-// GetProficiencyBonus returns the character's proficiency bonus based on level
-func (c *Character) GetProficiencyBonus() int {
-	if c.Level == 0 {
-		return 2 // Default for level 1
-	}
-	return 2 + ((c.Level - 1) / 4)
-}
-
-// HasSavingThrowProficiency checks if the character is proficient in a saving throw
-func (c *Character) HasSavingThrowProficiency(attribute Attribute) bool {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-
-	if c.Proficiencies == nil || c.Proficiencies[ProficiencyTypeSavingThrow] == nil {
-		return false
-	}
-
-	// Convert attribute to the expected saving throw key format
-	savingThrowKey := fmt.Sprintf("saving-throw-%s", strings.ToLower(string(attribute)))
-
-	for _, prof := range c.Proficiencies[ProficiencyTypeSavingThrow] {
-		if strings.EqualFold(prof.Key, savingThrowKey) {
-			return true
-		}
-	}
-
-	return false
-}
-
-// GetSavingThrowBonus calculates the total saving throw bonus for an attribute
-func (c *Character) GetSavingThrowBonus(attribute Attribute) int {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-
-	// Get ability modifier
-	modifier := 0
-	if abilityScore, exists := c.Attributes[attribute]; exists && abilityScore != nil {
-		modifier = abilityScore.Bonus
-	}
-
-	// Check proficiency without locking again
-	isProficient := false
-	if c.Proficiencies != nil && c.Proficiencies[ProficiencyTypeSavingThrow] != nil {
-		savingThrowKey := fmt.Sprintf("saving-throw-%s", strings.ToLower(string(attribute)))
-		for _, prof := range c.Proficiencies[ProficiencyTypeSavingThrow] {
-			if strings.EqualFold(prof.Key, savingThrowKey) {
-				isProficient = true
-				break
-			}
-		}
-	}
-
-	// Add proficiency bonus if proficient
-	if isProficient {
-		modifier += c.GetProficiencyBonus()
-	}
-
-	return modifier
-}
-
-// RollSavingThrow rolls a saving throw for the given attribute
-func (c *Character) RollSavingThrow(attribute Attribute) (*dice.RollResult, int, error) {
-	bonus := c.GetSavingThrowBonus(attribute)
-
-	// Roll 1d20
-	result, err := c.getDiceRoller().Roll(1, 20, bonus)
-	if err != nil {
-		return nil, 0, fmt.Errorf("failed to roll saving throw: %w", err)
-	}
-
-	return result, result.Total, nil
-}
-
-// HasSkillProficiency checks if the character is proficient in a skill
-func (c *Character) HasSkillProficiency(skillKey string) bool {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-
-	if c.Proficiencies == nil || c.Proficiencies[ProficiencyTypeSkill] == nil {
-		return false
-	}
-
-	for _, prof := range c.Proficiencies[ProficiencyTypeSkill] {
-		if strings.EqualFold(prof.Key, skillKey) {
-			return true
-		}
-	}
-
-	return false
-}
-
 // GetSkillBonus calculates the total skill bonus
 func (c *Character) GetSkillBonus(skillKey string, attribute Attribute) int {
 	c.mu.Lock()
@@ -1314,237 +742,6 @@ func (c *Character) RollSkillCheck(skillKey string, attribute Attribute) (*dice.
 	return result, result.Total, nil
 }
 
-// GetEffectManager returns the character's effect manager, initializing it if needed
-func (c *Character) GetEffectManager() *effects.Manager {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-
-	return c.getEffectManagerInternal()
-}
-
-// getEffectManagerInternal returns the effect manager without locking (caller must hold lock)
-func (c *Character) getEffectManagerInternal() *effects.Manager {
-	if c.EffectManager == nil {
-		c.EffectManager = effects.NewManager()
-		// Restore effects from persisted data
-		c.syncResourcestoEffectManager()
-	}
-	return c.EffectManager
-}
-
-// syncResourcestoEffectManager restores EffectManager from persisted ActiveEffects
-func (c *Character) syncResourcestoEffectManager() {
-	if c.Resources == nil || c.Resources.ActiveEffects == nil {
-		return
-	}
-
-	// Convert old ActiveEffect format back to new StatusEffect format
-	for _, oldEffect := range c.Resources.ActiveEffects {
-		if oldEffect == nil {
-			continue
-		}
-
-		// Create new status effect
-		newEffect := &effects.StatusEffect{
-			ID:          oldEffect.ID,
-			Name:        oldEffect.Name,
-			Description: oldEffect.Description,
-			Source:      effects.EffectSource(oldEffect.Source),
-			SourceID:    oldEffect.SourceID,
-			Duration: effects.Duration{
-				Rounds: oldEffect.Duration,
-			},
-			Modifiers:  []effects.Modifier{}, // TODO: Convert modifiers if needed
-			Conditions: []effects.Condition{},
-			Active:     true,
-		}
-
-		// Set duration type
-		switch oldEffect.DurationType {
-		case DurationTypePermanent:
-			newEffect.Duration.Type = effects.DurationPermanent
-		case DurationTypeRounds:
-			newEffect.Duration.Type = effects.DurationRounds
-		case DurationTypeUntilRest:
-			newEffect.Duration.Type = effects.DurationUntilRest
-		default:
-			newEffect.Duration.Type = effects.DurationPermanent
-		}
-
-		// For well-known effects, rebuild them properly
-		if oldEffect.Name == "Rage" && oldEffect.Source == string(effects.SourceAbility) {
-			// Rebuild rage effect with proper modifiers
-			rageEffect := effects.BuildRageEffect(c.Level)
-			rageEffect.ID = oldEffect.ID // Keep the same ID
-			if err := c.EffectManager.AddEffect(rageEffect); err != nil {
-				log.Printf("Failed to add rage effect: %v", err)
-			}
-		} else if oldEffect.Name == "Favored Enemy" && oldEffect.Source == string(effects.SourceFeature) {
-			// Rebuild favored enemy effect with the selected enemy type
-			enemyType := "humanoids" // default fallback
-			for _, feature := range c.Features {
-				if feature.Key == "favored_enemy" && feature.Metadata != nil {
-					if et, ok := feature.Metadata["enemy_type"].(string); ok {
-						enemyType = et
-						break
-					}
-				}
-			}
-			favoredEnemyEffect := effects.BuildFavoredEnemyEffect(enemyType)
-			favoredEnemyEffect.ID = oldEffect.ID
-			if err := c.EffectManager.AddEffect(favoredEnemyEffect); err != nil {
-				log.Printf("Failed to add favored enemy effect: %v", err)
-			}
-		} else {
-			// Add generic effect
-			if err := c.EffectManager.AddEffect(newEffect); err != nil {
-				log.Printf("Failed to add effect %s: %v", newEffect.Name, err)
-			}
-		}
-	}
-}
-
-// AddStatusEffect adds a new status effect to the character
-func (c *Character) AddStatusEffect(effect *effects.StatusEffect) error {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-
-	return c.addStatusEffectInternal(effect)
-}
-
-// addStatusEffectInternal adds a status effect without locking (caller must hold lock)
-func (c *Character) addStatusEffectInternal(effect *effects.StatusEffect) error {
-	err := c.getEffectManagerInternal().AddEffect(effect)
-	if err != nil {
-		return err
-	}
-
-	// Also sync to the persisted ActiveEffects for backward compatibility
-	c.syncEffectManagerToResources()
-	return nil
-}
-
-// syncEffectManagerToResources syncs EffectManager to the persisted ActiveEffects
-func (c *Character) syncEffectManagerToResources() {
-	if c.EffectManager == nil || c.Resources == nil {
-		return
-	}
-
-	// Get active effects from manager
-	activeEffects := c.EffectManager.GetActiveEffects()
-
-	// Clear old active effects
-	c.Resources.ActiveEffects = []*ActiveEffect{}
-
-	// Convert new status effects to old ActiveEffect format for persistence
-	for _, effect := range activeEffects {
-		if effect == nil {
-			continue
-		}
-
-		// Convert to old format (simplified for now)
-		oldEffect := &ActiveEffect{
-			ID:           effect.ID,
-			Name:         effect.Name,
-			Description:  effect.Description,
-			Source:       string(effect.Source),
-			SourceID:     effect.SourceID,
-			Duration:     effect.Duration.Rounds,
-			DurationType: DurationTypeRounds, // Simplified
-			Modifiers:    []Modifier{},       // TODO: Convert modifiers if needed
-		}
-
-		// Set duration type based on new system
-		switch effect.Duration.Type {
-		case effects.DurationPermanent:
-			oldEffect.DurationType = DurationTypePermanent
-		case effects.DurationRounds:
-			oldEffect.DurationType = DurationTypeRounds
-		case effects.DurationInstant:
-			oldEffect.DurationType = DurationTypeRounds // Instant effects map to rounds
-		case effects.DurationWhileEquipped:
-			oldEffect.DurationType = DurationTypePermanent
-		case effects.DurationUntilRest:
-			oldEffect.DurationType = DurationTypeUntilRest
-		}
-
-		c.Resources.ActiveEffects = append(c.Resources.ActiveEffects, oldEffect)
-	}
-}
-
-// RemoveStatusEffect removes a status effect by ID
-func (c *Character) RemoveStatusEffect(effectID string) {
-	c.GetEffectManager().RemoveEffect(effectID)
-}
-
-// GetActiveStatusEffects returns all active status effects
-func (c *Character) GetActiveStatusEffects() []*effects.StatusEffect {
-	return c.GetEffectManager().GetActiveEffects()
-}
-
-// GetAttackModifiers returns all modifiers that apply to attack rolls
-func (c *Character) GetAttackModifiers(conditions map[string]string) []effects.Modifier {
-	return c.GetEffectManager().GetModifiers(effects.TargetAttackRoll, conditions)
-}
-
-// GetDamageModifiers returns all modifiers that apply to damage rolls
-func (c *Character) GetDamageModifiers(conditions map[string]string) []effects.Modifier {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-
-	return c.getDamageModifiersInternal(conditions)
-}
-
-// getDamageModifiersInternal returns damage modifiers without locking (caller must hold lock)
-func (c *Character) getDamageModifiersInternal(conditions map[string]string) []effects.Modifier {
-	return c.getEffectManagerInternal().GetModifiers(effects.TargetDamage, conditions)
-}
-
-// ApplyDamageResistance applies resistance/vulnerability/immunity from status effects
-func (c *Character) ApplyDamageResistance(damageType damage.Type, amount int) int {
-	conditions := map[string]string{
-		"damage_type": string(damageType),
-	}
-
-	// Check for immunities
-	immunities := c.GetEffectManager().GetModifiers(effects.TargetImmunity, conditions)
-	for _, mod := range immunities {
-		if mod.DamageType == string(damageType) {
-			return 0 // Immune to this damage type
-		}
-	}
-
-	// Check for resistances
-	resistances := c.GetEffectManager().GetModifiers(effects.TargetResistance, conditions)
-	hasResistance := false
-	for _, mod := range resistances {
-		if mod.DamageType == string(damageType) {
-			hasResistance = true
-			break
-		}
-	}
-
-	// Check for vulnerabilities
-	vulnerabilities := c.GetEffectManager().GetModifiers(effects.TargetVulnerability, conditions)
-	hasVulnerability := false
-	for _, mod := range vulnerabilities {
-		if mod.DamageType == string(damageType) {
-			hasVulnerability = true
-			break
-		}
-	}
-
-	// Apply modifiers
-	if hasVulnerability {
-		amount *= 2
-	}
-	if hasResistance {
-		amount /= 2
-	}
-
-	return amount
-}
-
 // applyFightingStyleBonuses applies bonuses from fighter fighting styles
 func (c *Character) applyFightingStyleBonuses(weapon *Weapon, attackBonus, damageBonus int) (finalAttackBonus, finalDamageBonus int) {
 	return c.applyFightingStyleBonusesWithHand(weapon, attackBonus, damageBonus, SlotMainHand)
@@ -1576,7 +773,7 @@ func (c *Character) applyFightingStyleBonusesWithHand(weapon *Weapon, attackBonu
 
 	// Apply fighting style bonuses
 	switch fightingStyle {
-	case "archery":
+	case "archery": // TODO: these keys should be a typed constant
 		// +2 to attack rolls with ranged weapons
 		if weapon.IsRanged() {
 			attackBonus += 2
